added black boxes behind text

+++ createWatermark.py
@@ -1,4 +1,5 @@
 import configparser
+import os
 from shutil import copyfile
 from PIL import Image, ImageDraw, ImageFont
 from wmDataPrep import wmDictCreator
@@ -6,7 +7,7 @@ from wmDataPrep import wmDictCreator
 
 def createWatermark(filename, position, watermark, xmod, ymod):
     text = watermark
-    color = 'black'
+    color = 'white'
     fontfamily = 'arial.ttf'
     image = Image.open(config['PATHS']['imgout'] + filename).convert('RGBA')
     imageWatermark = Image.new('RGBA', image.size, (255, 255, 255, 0))
@@ -17,13 +18,15 @@ def createWatermark(filename, position, watermark, xmod, ymod):
     if position == 'bottom':
         x = (width - textWidth)/2
         y = height * ymod
+        box = [x, y, (x+textWidth), (y+textHeight)]
     else:
         x = width * xmod
         y = height * ymod
+        box = [x, y, (x+textWidth), (y+textHeight)]
+    draw.rectangle(box, fill='black', outline='black', width=0)
     draw.text((x, y), text, color, font)
     wmImg = Image.alpha_composite(image, imageWatermark)
     wmImg.convert('RGB').save(config['PATHS']['imgOut'] + filename)
-    return f"{filename}\t{watermark}"
 
 
 config = configparser.ConfigParser()
@@ -31,7 +34,12 @@ config.read('./configs/config.ini')
 
 files = wmDictCreator(config['PATHS']['incsv'])
 
+problemfiles = open(config['PATHS']['problemfiles'], 'w')
+
 for k, v in files.items():
-    copyfile(config['PATHS']['imgin'] + k, config['PATHS']['imgout'] + k)
-    for key, val in files[k].items():
-        createWatermark(k, key, val[0], val[1], val[2])
+    if os.path.isfile(config['PATHS']['imgin'] + k):
+        copyfile(config['PATHS']['imgin'] + k, config['PATHS']['imgout'] + k)
+        for key, val in files[k].items():
+            createWatermark(k, key, val[0], val[1], val[2])
+    else:
+        problemfiles.write(f'{k}\n')
